all binary ops and primitive test passing
